update :: forgot password removed, register to login route
